Secrets migrate: split plan/apply/backup modules
